feat(ui): refactor robust and adaptive Button component with modular styles
- Refactored Button component to support all usage contexts (labels, objects, JSX logs, etc.)
- Ensured compatibility with HTMLAnchorElement via `href` and improved props safety
- Added `aria-pressed` for better accessibility when `selected` is true
- Split button styling into modular CSS files: core, layouts, animations, and variant styles
- Applied BEM, OOCSS and SMACSS principles for scalability and maintainability
- Introduced `ButtonContent` and `generateButtonClassNames` helpers (optionally usable)
- Preserved compatibility with existing Storybook stories
