+++ apps/ui/src/components/button/helpers.tsx
@@ -1,15 +1,21 @@
-import type { Properties } from './types';
+import clsx from 'clsx';
+import type { ButtonProperties } from './types';
 
-/**
- *
- * @param properties
- * @returns
- */
-// after
-export function toDefaults(properties?: Properties): Required<Properties> {
-	return {
-		as: 'button',
-		children: properties?.children ?? null,
-		...properties,
-	} as Required<Properties>;
+export function generateButtonClassNames({
+	variant,
+	fullWidth,
+	selected,
+	layout,
+	className,
+}: Partial<ButtonProperties>) {
+	return clsx(
+		'button',
+		variant && `button--${variant}`,
+		{
+			'button--fullWidth': fullWidth,
+			'button--selected': selected,
+		},
+		layout && `layout--${layout}`,
+		className
+	);
 }
