+++ apps/ui/src/components/button/index.ts
@@ -1,3 +1,2 @@
-export { default as Button } from './button';
-
-export type { Properties as ButtonProperties } from './types';
+export { default } from './button';
+export * from './types';
