+++ apps/ui/src/components/button/styles/variables.css
@@ -1,46 +1,42 @@
 :root {
-  /* === COLOR TOKENS === */
-  /* Filled Button */
-  --gol-btn-filled-bg: #6750a4;
-  --gol-btn-filled-text: #ffffff;
-  --gol-btn-filled-hover: #5a4795;
+  /* === COLOR: Surface & States === */
+  --md-sys-color-surface: #ffffff;
+  --md-sys-color-on-surface: #1d1b20;
+  --md-sys-color-surface-variant: #e7e0ec;
+  --md-sys-color-on-surface-variant: #49454f;
+  --md-sys-color-outline: #79747e;
 
-  /* Elevated Button */
-  --gol-btn-elevated-bg: #ffffff;
-  --gol-btn-elevated-text: #1d1b20;
-  --gol-btn-elevated-hover: #f1edf5;
+  --md-state-hover-layer: rgba(0, 0, 0, 0.08);
+  --md-state-pressed-layer: rgba(0, 0, 0, 0.12);
+  --md-disabled-content: rgba(0, 0, 0, 0.38);
+  --md-disabled-bg: rgba(29, 27, 32, 0.12);
 
-  /* Tonal Button */
-  --gol-btn-tonal-bg: #625b71;
-  --gol-btn-tonal-text: #ffffff;
-  --gol-btn-tonal-hover: #574f65;
-
-  /* Outlined Button */
-  --gol-btn-outlined-border: #79747e;
-  --gol-btn-outlined-text: #1d1b20;
-  --gol-btn-outlined-hover: rgba(0, 0, 0, 0.05);
-
-  /* Text Button */
-  --gol-btn-text-text: #1d1b20;
-  --gol-btn-text-hover: rgba(0, 0, 0, 0.05);
-
-  /* === ELEVATION TOKENS === */
-  --gol-elevation-1: 0px 1px 3px rgba(0, 0, 0, 0.2);
-  --gol-elevation-2: 0px 3px 6px rgba(0, 0, 0, 0.3);
-  --gol-elevation-3: 0px 6px 10px rgba(0, 0, 0, 0.4);
-
-  /* === TRANSITIONS === */
-  --gol-transition-fast: 0.15s ease-in-out;
-  --gol-transition-base: 0.2s ease-in-out;
-  --gol-transition-slow: 0.3s ease-in-out;
+  /* === COLOR: Primary === */
+  --md-sys-color-primary: #6750a4;
+  --md-sys-color-on-primary: #ffffff;
+  --md-sys-color-primary-container: #eaddff;
+  --md-sys-color-on-primary-container: #21005d;
 
   /* === TYPOGRAPHY === */
-  --gol-font-family-base: 'Roboto', sans-serif;
-  --gol-font-size-base: 14px;
+  --md-typescale-label-large-family: 'Roboto', sans-serif;
+  --md-typescale-label-large-size: 14px;
 
-  /* === SPACING & RADIUS === */
-  --gol-btn-padding-y: 8px;
-  --gol-btn-padding-x: 16px;
-  --gol-btn-gap: 8px;
-  --gol-radius-sm: 4px;
+  /* === ELEVATION === */
+  --md-elevation-0: none;
+  --md-elevation-1: 0px 1px 3px rgba(0, 0, 0, 0.2);
+  --md-elevation-2: 0px 3px 6px rgba(0, 0, 0, 0.3);
+
+  /* === SPACING === */
+  --md-spacing-button-padding-y: 8px;
+  --md-spacing-button-padding-x: 16px;
+  --md-spacing-button-gap: 8px;
+  --md-radius-button: 4px;
+
+  /* === FOCUS === */
+  --md-focus-ring: #6750a4;
+
+  /* === TRANSITIONS === */
+  --md-transition-fast: 0.15s ease-in-out;
+  --md-transition-base: 0.2s ease-in-out;
+  --md-transition-slow: 0.3s ease-in-out;
 }
