+++ apps/ui/src/components/button/styles/variants/filled.css
@@ -1,32 +1,26 @@
 @import url('../variables.css');
 @import url('../animations.css');
 
-/* === BUTTON: Filled Variant === */
-
 .button--filled {
-  background-color: var(--md-filled-bg);
-  color: var(--md-filled-text);
+  background-color: var(--md-sys-color-primary);
+  color: var(--md-sys-color-on-primary);
   box-shadow: var(--md-elevation-1);
-  transition: background-color 0.2s ease, box-shadow 0.2s ease;
+  border: none;
+  transition: background-color var(--md-transition-base), box-shadow var(--md-transition-base);
 }
 
-/* === Hover State === */
 .button--filled:hover {
-  background-color: var(--md-filled-hover);
+  background-color: #5a4795; 
   box-shadow: var(--md-elevation-2);
 }
 
-/* === Pressed State === */
 .button--filled:active {
-  animation: gol-button-press 0.2s ease-in-out;
+  animation: gol-button-press var(--md-transition-base);
   box-shadow: var(--md-elevation-1);
 }
 
-/* === Disabled State === */
 .button--filled:disabled {
-  background-color: rgba(103, 80, 164, 0.12); /* On-surface overlay */
-  color: rgba(0, 0, 0, 0.38);
+  background-color: var(--md-disabled-bg);
+  color: var(--md-disabled-content);
   box-shadow: none;
-  cursor: default;
-  pointer-events: none;
 }
