+++ apps/ui/src/components/button/styles/variants/outlined.css
@@ -1,30 +1,22 @@
 @import url('../variables.css');
 @import url('../animations.css');
 
-/* === BUTTON: Outlined Variant === */
-
 .button--outlined {
   background-color: transparent;
-  color: var(--md-outlined-text);
-  border: 1px solid var(--md-outlined-border);
-  transition: background-color 0.2s ease, border-color 0.2s ease;
+  color: var(--md-sys-color-primary);
+  border: 1px solid var(--md-sys-color-outline);
+  transition: background-color var(--md-transition-base), border-color var(--md-transition-base);
 }
 
-/* === Hover State === */
 .button--outlined:hover {
-  background-color: var(--md-outlined-hover);
-  border-color: var(--md-outlined-border);
+  background-color: var(--md-state-hover-layer);
 }
 
-/* === Pressed State === */
 .button--outlined:active {
-  animation: gol-button-press 0.2s ease-in-out;
+  animation: gol-button-press var(--md-transition-base);
 }
 
-/* === Disabled State === */
 .button--outlined:disabled {
-  border-color: rgba(29, 27, 32, 0.12);
-  color: rgba(0, 0, 0, 0.38);
-  cursor: default;
-  pointer-events: none;
+  border-color: var(--md-disabled-bg);
+  color: var(--md-disabled-content);
 }
